fix few things

+++ src/app/dashboard/inventory/product-overview/product-overview.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit, computed, signal } from '@angular/core';
 import { ProductService } from '../../../services/product.service';
 import { Product } from '../../../models/products.models';
 import { ToastService } from '../../../services/toast.service';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'app-product-overview',
@@ -15,7 +16,7 @@ export class ProductOverviewComponent implements OnInit {
 
 
 
-  constructor (private productsService: ProductService, private toast: ToastService) {
+  constructor (private productsService: ProductService, private toast: ToastService, private connfirmation: ConfirmationService) {
   }
 
   ngOnInit(): void {
@@ -32,16 +33,24 @@ export class ProductOverviewComponent implements OnInit {
   }
 
   deleteProduct(id: string) {
-    this.loading.set(true)
-    this.productsService.deleteStoreProduct(id).subscribe({
-      next: () => {
-        this.toast.success('Product Deleted')
-        this.loading.set(false)
+    this.connfirmation.confirm({
+      header: 'Delete Product',
+      message: 'Are you sure you want to delete this product?',
+      accept: () => {
+        this.loading.set(true);
+        this.productsService.deleteStoreProduct(id).subscribe({
+          next: () => {
+            this.toast.success('Product Deleted');
+            this.loading.set(false);
+          },
+          error: (err) => {
+            this.toast.error(err);
+            this.loading.set(false);
+          },
+        });
       },
-      error: (err) => {
-        this.toast.error(err)
-        this.loading.set(false)
-      }
+      reject: () => {}
     })
+
   }
 }
